feat: change -rtorrent.collect.active to -rtorrent.downloads.collect.details
Also adds a new structure: CollectorOpts which will allow us to add more
parameters in the future without having to constantly expand the
function parameters list over and over in the future.

Also refactors the way that we parse the XMLRPC communication to be a
bit more resilient and understandable. Rather than relying on strict
array placement, we now process the slice in the same way that the
commands are specified. This also allows us to add more commands in the
future if we want to.