fix: fundamental misunderstanding of the rtorrent API
Before I thought that multiple URLs for trackers could exist on a
Tracker object, however, after running it through more examples, I
that if a tracker has multiple URLs, that it instead returns

This caused me to need to rework the library somewhat significantly.

I also added some additional logging and some updating of the caching
library to ensure that it is a bit more robust.